+++ passes/S040/testdata/src/a/comment_ignore_v2.go
@@ -1,0 +1,14 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+func fcommentignore_v2() {
+	//lintignore:S040
+	_ = schema.Schema{
+		Computed:         true,
+		Type:             schema.TypeString,
+		ValidateDiagFunc: validateDiagFunc_v2,
+	}
+}