+++ passes/S043/testdata/src/a/comment_ignore_v2.go
@@ -1,0 +1,22 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+func fcommentignore_v2() {
+	//lintignore:S043
+	_ = schema.Schema{
+		Type:     schema.TypeSet,
+		Optional: true,
+		Elem: &schema.Resource{
+			Schema: map[string]*schema.Schema{
+				"token": {
+					Type:      schema.TypeString,
+					Optional:  true,
+					WriteOnly: true,
+				},
+			},
+		},
+	}
+}